+++ typescript-sdk/jsr.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://jsr.io/schema/config-file.v1.json",
   "name": "@union/client",
-  "version": "0.0.34",
+  "version": "0.0.35",
   "license": "MIT",
   "exports": {
     ".": "./src/mod.ts"
