+++ typescript-sdk/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@unionlabs/client",
-  "version": "0.0.34",
+  "version": "0.0.35",
   "homepage": "https://union.build",
   "description": "Union Labs cross-chain transfers client",
   "type": "module",
@@ -49,10 +49,10 @@
     "cosmjs-types": "^0.9.0",
     "jsr": "^0.13.2",
     "tsup": "^8.3.5",
-    "tsx": "^4.19.1",
+    "tsx": "^4.19.2",
     "typescript": "^5.6.3",
     "vite-tsconfig-paths": "^5.0.1",
-    "vitest": "^2.1.3"
+    "vitest": "^2.1.4"
   },
   "repository": {
     "type": "git",
