+++ typescript-sdk/src/client/evm.ts
@@ -204,4 +204,4 @@ export const createEvmClient = (parameters: EvmClientParameters) => {
         })
       }
     }))
-}
+}
