+++ typescript-sdk/src/client/types.ts
@@ -1,10 +1,4 @@
-import type {
-  aptosChainId,
-  AptosChainId,
-  AptosAccount,
-  AptosBrowserWallet,
-  AptosPublicAccountInfo
-} from "./aptos.ts"
+import type { aptosChainId, AptosChainId, AptosAccount } from "../aptos/client.ts"
 import type { OfflineSigner } from "../types.ts"
 import type { evmChainId, EvmChainId } from "./evm.ts"
 import type { Account as ViemAccount, Address } from "viem"
@@ -42,17 +36,8 @@ export type TransferAssetsParameters<CHAIN_ID extends EvmChainId | CosmosChainId
           simulate?: boolean
           denomAddress: string
           account?: AptosAccount
+          authAccess: "key" | "wallet"
           relayContractAddress?: string
           gasPrice?: { amount: string; denom: string }
-        } & (
-          | {
-              authAccess: "key"
-              account?: AptosAccount
-            }
-          | {
-              authAccess: "wallet"
-              account?: AptosPublicAccountInfo
-              sign: AptosBrowserWallet["signTransaction"]
-            }
-        )
+        }
       : undefined)
